refactor: split module
Move io logic for text-files and directories into separate modules
